feat. 재고 폐기 구현
- 안전재고량 확인
- 폐기 시 안전재고량을 만족하는 재고되도록 구현

+++ src/main/resources/static/js/inventory/adjust.js
@@ -87,7 +87,7 @@ function updateSector() {
                 return;
             }
             idList.push(inventoryId);
-            sectorList.push(realQuantity);
+            sectorList.push(sectorName);
         }
     });
 
@@ -114,8 +114,16 @@ function deleteInventory() {
         }
     });
 
+    if (idList.length === 0) {
+        alert("삭제할 재고가 없습니다.");
+        return;
+    }
+
+    console.log(idList);
     axios.delete("/inventory/deleteInventory", {
-        idList: idList
+        data: {
+            idList: idList
+        }
     }).then(response => {
         if (response.data > 0) {
             location.reload();
